ml-service/requirements.txt
fastapi[all]
uvicorn
python-multipart
torch
torchvision
Pillow
numpy
